Working Linked List.
Linked list takes in new data and inserts it into the list in
alphabetical order. If the list item is already in the list it
increments the counter of the node that already holds that word. The
displayForward function works mostly. It's first output is some NULL
data. Need to fix that.